requirements.txt
numpy==1.26.4
scipy==1.12.0
cvxpy==1.4.2
PyYAML==6.0.2
